Update README and finalize features (Run Menu, Themes, Backup)

+++ scintilla/include/Sci_Position.h
@@ -10,7 +10,7 @@
 #define SCI_POSITION_H
 
 #include <stddef.h>
-#include <cstdint>
+#include "ScintillaTypes.h"
 
 // Basic signed type used throughout interface
 typedef ptrdiff_t Sci_Position;
@@ -19,7 +19,7 @@ typedef ptrdiff_t Sci_Position;
 typedef size_t Sci_PositionU;
 
 // For Sci_CharacterRange  which is defined as long to be compatible with Win32 CHARRANGE
-typedef intptr_t Sci_PositionCR;
+typedef long Sci_PositionCR;
 
 #ifdef _WIN32
 	#define SCI_METHOD __stdcall
